# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "API documentation description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/admin/types": {
            "get": {
                "tags": [
                    "Types"
                ],
                "summary": "Get list of types",
                "description": "Display a listing of the resource.",
                "operationId": "6985064bf1d252f4d9cbb259ea74cf4e",
                "parameters": [
                    {
                        "name": "bulk",
                        "in": "query",
                        "description": "Return list of IDs only",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of types",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Type"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Types"
                ],
                "summary": "Create a new type",
                "description": "Store a newly created resource in storage.",
                "operationId": "34ad3424b448f6fe6f3be6e93a2ba181",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Type"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Type created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "redirect": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/types/{id}": {
            "put": {
                "tags": [
                    "Types"
                ],
                "summary": "Update a specific type",
                "description": "Update the specified resource in storage.",
                "operationId": "3e9804cbe923843f18c092b4618ee933",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Type ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Type"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Type updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "redirect": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Types"
                ],
                "summary": "Delete a specific type",
                "description": "Remove the specified resource from storage.",
                "operationId": "2cc4b5c64fd332e17e7a706917b328b1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Type ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Type deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/types/bulk-destroy": {
            "post": {
                "tags": [
                    "Types"
                ],
                "summary": "Bulk delete types",
                "description": "Remove the specified resources from storage.",
                "operationId": "d8512b13039dbdbe55cd08ebc1023735",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Types deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Type": {
                "title": "Type",
                "required": [
                    "slug",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "Type ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "slug": {
                        "description": "Unique slug identifier",
                        "type": "string"
                    },
                    "name": {
                        "description": "Display name of the type (translatable)",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "resource_url": {
                        "description": "URL to the admin interface for this type",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}